feat: handle insert from oplog
